Fix multiple transformers fighting for inputStack, setting the NBT to null, and causing a NPE later on
Whenever two (or more) transformers would fulfil their conditions at once, the first one would correctly pass the tank's fluid over to inputStack and remove the fluid, but the second transformer would retrieve the null fluid and assign it to inputStack.
This would set the NBT data to null, causing a NPE both in the barrel's render method due to a null stack, and in the update method due to passing null to the transformer.